move schema silver para raiz
