+++ etl/tests/test_schema.py
@@ -1,6 +1,6 @@
 from django.test import SimpleTestCase
 
-from etl.indexing.schema import SILVER_MAPPING, SILVER_PROPERTIES
+from etl.schema import SILVER_MAPPING, SILVER_PROPERTIES
 
 
 class SilverMappingTests(SimpleTestCase):
